Add support for COS and Cray X-MP

+++ h/cgg_cg.h
@@ -80,8 +80,8 @@ typedef struct {
 
 struct exprnode {
 	short ex_operator;
-	short ex_lnode;
-	short ex_rnode;
+	long  ex_lnode;
+	long  ex_rnode;
 };
 typedef struct exprnode node_t;
 typedef struct exprnode *node_p;
